backend/requirements.txt
Flask
Flask-RESTful
Flask-Cors
werkzeug
PyMySQL
flask-sqlalchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
mysql-connector-python
cryptography